admin can add users by sending email invitation

+++ routes/web.php
@@ -11,9 +11,7 @@
 |
 */
 
-Route::get('/', function () {
-    return view('customAuth.login');
-});
+Route::get('/', 'CustomAuth\LoginController@showLoginForm')->name('clogin');
 
 Auth::routes();
 
@@ -21,28 +19,29 @@ Auth::routes();
 Route::get('/cregister', 'CustomAuth\RegisterController@showRegistrationForm');
 Route::post('/cregister', 'CustomAuth\RegisterController@register');
 
-Route::get('/clogin', 'CustomAuth\LoginController@showLoginForm');
+Route::get('/clogin', 'CustomAuth\LoginController@showLoginForm')->name('clogin');
 Route::post('/clogin', 'CustomAuth\LoginController@authenticate');
 
-
+Route::get('/invitation', 'CustomAuth\ActivateController@showActivationForm');
+Route::post('/invitation', 'CustomAuth\ActivateController@activate');
 
 Route::get('/home', 'HomeController@index')->name('home')->middleware('auth');
 
 //Route::resource('/admin/users','Admin\UsersController',['except'=>['show','create','store']]);
 
 Route::namespace('Admin')->prefix('admin')->name('admin.')->middleware('can:manage-users')->group(function(){
-    Route::resource('/users','UsersController',['except'=>['show','create','store']]);
+    Route::resource('/users','UsersController',['except'=>['show']]);
 });
 
 
 
-Route::get('/clients', 'ClientsController@index');
-Route::post('/clients', 'ClientsController@store');
-Route::get('/clients/create', 'ClientsController@create');
-Route::get('/clients/{client}/edit', 'ClientsController@edit');
-Route::put('/clients/{client}', 'ClientsController@update');
+Route::get('/clients', 'ClientsController@index')->middleware('can:is_admin');
+Route::post('/clients', 'ClientsController@store')->middleware('can:is_admin');
+Route::get('/clients/create', 'ClientsController@create')->middleware('can:is_admin');
+Route::get('/clients/{client}/edit', 'ClientsController@edit')->middleware('can:is_admin');
+Route::put('/clients/{client}', 'ClientsController@update')->middleware('can:is_admin');
 //Route::delete('/clients/{client}', 'ClientsController@delete');
-Route::delete('/clients/{client}', 'ClientsController@ajaxDelete');
+Route::delete('/clients/{client}', 'ClientsController@ajaxDelete')->middleware('can:is_admin');
 
 Route::get('/tasks', 'TasksController@index');
 Route::post('/tasks', 'TasksController@store');
